models/my-model/requirements.txt
boto3
sagemaker
pandas
scikit-learn
pyyaml
flask
